Spaces are no longer valid
There is a discrepancy between what the RFC (or at least this gem)
considers to be a valid email address and what our Email provider
actually allows. Since our email provider does not allow leading or
trailing spaces, it seemed simpler to make all spaces invalid rather
than allow for slash escaping characters in the middle of the strings.